chnged type of hcll and lcll to int in OtherControls.h OtherControls.cpp and UI

+++ OtherControls.h
@@ -24,7 +24,7 @@ class OtherControl {
         OtherControl(); // Default Constructor
 
         bool createReservation(int phoneNumber, string& sailingId, string& licensePlate); // Makes a new reservation, in: sailingId, licensePlate
-        bool createVessel(string& vesselName, double HCLL, double LCLL); //in: vesselName, HCLL, LCLL
+        bool createVessel(string& vesselName, int HCLL, int LCLL); //in: vesselName, HCLL, LCLL
         bool deleteReservation(string& sailingId); // Deletes all reservations for a sailing, in: sailingId
         bool deleteSailing(string& sailingId); // Deletes a sailing record, in: sailingId
         bool checkIn(string& licensePlate, std::string& sailingId); // Checks in a vehicle to a sailing, in: sailingId, licensePlate
